pyproject.toml
[project]
name = "stockevent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ics>=0.7.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
